Update to enable sessions
Sessions weren't working, and as a consequence
it was impossible to remove things from the
inventory.

This update enables sessions in myapp_local.wsgi
and myapp_nonlocal.wsgi. I also updated
cherrypy_local.conf and cherrypy_nonlocal.conf
to enable sessions. The .conf files SHOULD be
the only files I need to edit, but for some reason
that's not working. That's a bug I need to fix.