Fixed search bug. Add inventory now uses price from ISBN
I fixed the search bug where a search with an ISBN with a price field
was not working.

I also updated tools/inventory.py so that if, when adding a book
to inventory, the ISBN contains a price field that price is autopopulated
in the List Price and Our Price text boxes for the add inventory screen.